klip-mission-control/requirements.txt
fastapi>=0.104.0
python-multipart>=0.0.9
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis[hiredis]>=5.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
structlog>=23.0.0
httpx>=0.25.0
requests>=2.31.0
boto3>=1.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
